extract Class
